// src/app/Interfaces/prods_tipos.interface.ts
export interface TipoProducto {
    TIPO?:              string;
    correo:             string;
    correo2:            string;
    correo3:            string;
    correo4:            string;
    correo5:            string;
    alias:              string;
    paraPedido?:        number;
    bloqueado?:         number;
    diasPedido?:        number;
    dias_pedido?:       number;
    pedido_temp?:       boolean;
    sucursal:           string;
    consec?:            number;
    id_consec_sucursal?:string;
    diasPedidoMaximo?:  number;
}